+++ apple-codesign/src/error.rs
@@ -381,4 +381,7 @@ pub enum AppleCodesignError {
 
     #[error("plist: {0}")]
     Plist(#[from] plist::Error),
+
+    #[error("config error: {0:?}")]
+    Figment(#[from] figment::Error),
 }