+++ apple-codesign/src/main.rs
@@ -2,9 +2,29 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+use apple_codesign::AppleCodesignError;
+
 fn main() {
     let exit_code = match apple_codesign::cli::main_impl() {
         Ok(()) => 0,
+        Err(AppleCodesignError::Figment(err)) => {
+            eprintln!("configuration file error");
+
+            err.metadata.as_ref().map(|metadata| {
+                metadata.source.as_ref().map(|source| {
+                    source.file_path().map(|path| {
+                        eprintln!("  source path: {}", path.display());
+                    })
+                })
+            });
+            err.profile.as_ref().map(|profile| {
+                eprintln!("  in profile: {}", profile);
+            });
+            eprintln!("  problem key: {}", err.path.join(", "));
+            eprintln!("  problem: {:?}", err.kind);
+
+            1
+        }
         Err(err) => {
             eprintln!("Error: {err}");
             1